It lives!
If you poke a little program into memory, starting at 0x00 (or anywhere
later since instruction 0 is NOP) it'll run!
now I have to make it load binary files.

There is an ultra-verbose output that will print the status as well as
the entirety of memory. I am going to make that more compact now.